+++ PHP/Controleur/Accueil/loadChat.php
@@ -1,7 +1,7 @@
 <?php
 session_start();
 /* Cette page permet au chat d'être tous le temps a jour, et d'afficher le bon chat, sur le bon groupe*/
-require("../../Modele/Fonction/ConnectionBDD.php");
+require("../../Modele/BDD/ConnectionBDD.php");
 $conn= ConnectionBDD::getInstance();
 $bdd=$conn::getpdo();
 $tous = $bdd->query("SELECT idgroupe FROM groupe");
