+++ PHP/Modele/BDD/ConnectionBDD.php
