+++ PHP/View/Accueil/CreationCompte.html
@@ -1,0 +1,86 @@
+
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Connexion</title>
+    <link rel="stylesheet" href="../../View/Style/SiteIFSI.css">
+    <script src="../../Modele/Fonction/LesFonctionsJS.js"></script>
+
+</head>
+
+<body>
+<!--Le haut de la page avec l'image et le titre-->
+
+<header>
+    <div class="traduction">
+        <form action="../../Controleur/Accueil/LoginAnglais.php" method="post">
+            <input type="submit" value="English">
+        </form>
+    </div>
+    <a href="Accueil.php">
+        <img src="../../View/image/logoIFSI.png" width=234 height=125 alt="" >
+    </a>
+    <h1> Institut de Formation aux Soins Infirmiers (IFSI)</h1>
+    <br><br>
+</header>
+
+<!--La fleche pour revenir sur la page accueil-->
+
+
+
+<!--La box de l'inscription-->
+
+<h2>INSCRIPTION IFSI</h2>
+
+<div class="inscription">
+    <form method="post">
+        <label>NOM</label>
+        <br>
+        <input type="text" name="nom" id="nom" value="<?php echo @$_POST['nom']?>"  required>
+        <br>
+        <label>PRENOM </label>
+        <br>
+        <input type="text" name="prenom" id="prenom" value="<?php echo @$_POST['prenom']?>" required>
+        <br>
+        <label>EMAIL</label>
+        <br>
+        <input type="email" name="mail" id="mail"  value="<?php echo @$_POST['mail']?>" required>
+        <br>
+        <label>CODE</label>
+        <br>
+        <!--le code peut être commun-->
+        <input type="text" name="code" id="code" value="<?php echo @$_POST['code']?>"  required>
+        <br>
+        <div>
+            <p><a href="#" class="MDP">MOT DE PASSE<span>&ensp;- min une majuscule &ensp;<br> &ensp;- min 8 caractères &ensp;<br>&ensp; - min un chiffre &ensp;</span></a>
+        </div>
+        <label>
+            <input type="password" name="mdp" id="mdp" required>
+        </label>
+        <button type="button" id="pass1" onclick="changer('mdp')">O</button>
+        <br>
+        <div>
+            <p><a href="#" class="MDP">CONFIRMATION<span>&ensp;- min une majuscule &ensp;<br> &ensp;- min 8 caractères &ensp;<br>&ensp; - min un chiffre &ensp;</span></a>
+        </div>
+        <label>
+            <input type="password" name="mdp2" id="mdp2" required>
+        </label>
+        <button type="button" id="pass2" onclick="changer('mdp2')">O</button>
+        <br>
+
+        <input type="submit" value="Valider" >
+    </form>
+    <br>
+
+</div>
+
+<br>
+<form action="BesoinAide.php" method="post">
+    <button style="margin-left: 39%" class="button-28"  type="submit">Besoin d'aide</button>
+</form>
+
+</body>
+</html>
+
