+++ PHP/View/Accueil/MDPoublie.html
@@ -1,0 +1,70 @@
+
+<!doctype html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>MDPoublié</title>
+    <link rel="stylesheet" href="../../View/Style/SiteIFSI.css">
+</head>
+<body>
+<!--Le haut de la page avec l'image et le titre-->
+
+<header>
+    <div class="traduction">
+        <form action="../../Controleur/Accueil/MotDePasseOublieAnglais.php" method="post">
+            <input type="submit" value="English">
+        </form>
+    </div>
+    <a href="Accueil.php">
+        <img src="../../View/image/logoIFSI.png" width=234 height=125 alt="" >
+    </a>
+    <h1> Institut de Formation aux Soins Infirmiers (IFSI)</h1>
+    <br><br>
+</header>
+
+
+
+
+
+<!--box milieu-->
+
+<h3>    Entrez votre Email, <br>
+    vous allez recevoir un mail pour vérifier votre identitée
+</h3>
+<br>
+
+<div class ="inscription">
+
+    <form action="../../Controleur/Accueil/MotDePasseOublie.php" method="post">
+
+        <br>
+        <label>Email :</label>
+        <br><br>
+        <input type="email" name='mail' placeholder="Entrez votre Email" required>
+        <br>
+        <br>
+        <label>Confirmez votre email :</label>
+        <br><br>
+        <input type="email" name='mailV2' placeholder="Confirmez votre Email" required>
+        <br>
+        <p><input type="submit" value="Valider"></p>
+    </form>
+    <br>
+
+</div>
+
+<!--Le bas de la page-->
+
+<footer>
+    <form action="../../Controleur/Accueil/Login.php" method="post">
+        <input type="submit" value="Créer un compte">
+    </form> <br>
+    <form action="BesoinAide.php" method="post">
+        <input type="submit" value="Besoin d'aide ?">
+    </form>
+</footer>
+
+
+</body>
+</html>
