+++ PHP/View/Accueil/nouveauMDPAnglais.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <title>NouveauMDP</title>
+  <link rel="stylesheet" href="../../View/Style/SiteIFSI.css">
+  <script src="../../Modele/Fonction/LesFonctionsJS.js"></script>
+</head>
+<body>
+<!--Le haut de la page avec l'image et le titre-->
+
+<header>
+  <div class="traduction">
+    <form action="../../Controleur/Accueil/nouveaumdp.php" method="post">
+      <input type="submit" value="Francais">
+    </form>
+  </div>
+  <a href="AccueilAnglais.php">
+    <img src="../../View/image/logoIFSI.png" width=234 height=125 alt="" >
+  </a>
+  <h1>  Institute of Nursing Training</h1>
+  <br><br>
+</header>
+
+
+<!--box de milieu-->
+
+<h3>
+  Enter your new password, <br>
+  Then go back to login.
+</h3>
+<br>
+
+<div class ="inscription">
+
+  <form method="post">
+    <br>
+    <label>New Password :</label>
+    <br><br>
+    <input type="password" name='mdp' id="mdp" placeholder="Enter your new password" required ><button type="button" onclick="changer('mdp')">O</button>
+    <br>
+    <br>
+    <label>Confirm your password  :</label>
+    <br><br>
+    <input type="password" name='mdp2' id="mdp2" placeholder="Confirm your password" required><button type="button" onclick="changer('mdp2')">O</button>
+    <br>
+    <p><input type="submit" value="Validate"></p>
+  </form>
+
+</div>
+
+
+<!--Le bas de la page-->
+
+<footer>
+  <form action="../../Controleur/Accueil/LoginAnglais.php" method="post">
+    <input type="submit" value="Create an account">
+  </form> <br>
+  <form action="../../Controleur/Accueil/AccueilAnglais.php" method="post">
+    <input type="submit" value="login">
+  </form> <br>
+  <form action="BesoinAide.php" method="post">
+    <input type="submit" value="Need Help ?">
+  </form>
+</footer>
+</body>
+
+</html>
