+++ PHP/View/Etudiant/PageEtu.php
@@ -6,25 +6,25 @@ session_start();
 
 <head>
     <meta charset="UTF-8">
-    <title>ConnexionProfesseur</title>
-    <link rel="stylesheet" href="../../View/Style/PageProf.css" >
+    <title>ConnexionEtudiant </title>
+    <link rel="stylesheet" href="../Style/PageProf.css" >
     <script src="../../Modele/Fonction/LesFonctionsJS.js"></script>
 
 </head>
 <body>
 <?php
-include("../../View/BarreHTML/BarreScenarioAccueil.php");
+include("../../View/BarreHTML/BarreScenarioAccueilEtu.php");
 ?>
+
 <div class="texte">
     Bienvenue,<br> sur le site de l'IFSI consacré à la création de scénario <br>
-    En tant que professeur, vous avez accès à la création de scénario, <br>
-    mais aussi aux corrections et notes. <br>
+    En tant qu'étudiant, vous avez accès aux exercices de scénarios, <br>
+    mais aussi aux corrections et aux notes envoyées par votre professeur. <br>
     Vous pouvez également consulter votre messagerie.
 
 
-
 </div>
+
 </body>
 </html>
-
